Fixed previous commit
We fix the previous commit where we tried to stop the game, and set the winning team, based on the game state.

This since our previous commit didn't fix the problem either.

Now we added print statements within the Wordle and Bingo round, and also set the winning team within the Wordle and Bingo round functions (found in lingo.py and wordle.py).