Directly set winning/losing team when it happens
Directly set the team that won/lose the Bingo or Wordle game within the round itself, instead of within the Lingo loop.

We do this to that we don't need to check multiple times if the team has for example a filled row on the Bingo board, or other win conditions. Now instead we just check if a boolean is true or false.

We also removed some functions, and test cases, since those aren't needed anymore (with how the program is structured right now).