Separated words utils and wordle utils
Separated the functionality of getting a random word, and the wordle utility functions itself.

We also added a separate test file for the now words_utils.py.

Do note that we now don't test if the wordle game doesn't use the same Wordle word twice. This is because the structure of the functions doesn't let us test it in a good way yet.

But since that wasn't a requirement of this assignment anyway, it doesn't matter that much if it does happen.